NotificationsToggle - use MiqIcon and carbon icons
